+++ EasyReflectometry/calculators/__init__.py
@@ -4,7 +4,7 @@ __version__ = '0.0.1'
 import traceback
 
 try:
-    from .refnx.interface import Refnx  # noqa: F401
+    from .refnx.calculators import Refnx  # noqa: F401
 except Exception:
     traceback.print_exc()
     print('Warning: refnx is not installed')
@@ -16,7 +16,7 @@ except Exception:
 #     print('Warning: BornAgain python is not installed')
 
 try:
-    from .refl1d.interface import Refl1d  # noqa: F401
+    from .refl1d.calculators import Refl1d  # noqa: F401
 except Exception:
     traceback.print_exc()
     print('Warning: refl1d is not installed')
