+++ tests/calculators/refnx/test_refnx_calculator.py
@@ -8,7 +8,7 @@ import os
 import unittest
 import numpy as np
 from numpy.testing import assert_almost_equal, assert_equal
-from EasyReflectometry.interfaces.refnx.interface import Refnx
+from EasyReflectometry.calculators.refnx.calculators import Refnx
 from EasyReflectometry.sample.material import Material
